Changed way of adding and removing channels

+++ src/setupReleasesChannel.js
@@ -1,20 +1,25 @@
 const db = require('../api/mongoDB-funcs');
+const {releasesChannels} = require('../api/discord-properties');
 
 async function addChannel(msgDiscord) {
     const hasPermission = msgDiscord.member.roles.cache.some(role => role.name === 'New Releases Manager');
 
     if(hasPermission){
-        const numberReleasesChannels = await db.numberReleasesChannels(msgDiscord.guild.id);
+        let idReleasesChannels = releasesChannels.get(msgDiscord.guild.id);
 
-        if(numberReleasesChannels < 10){
-            const channelAdded = await db.addReleasesChannel(msgDiscord.channel.id, msgDiscord.guild.id);
-            if(channelAdded){
+        if(!idReleasesChannels.includes(msgDiscord.channel.id)){
+            if(idReleasesChannels.length < 10){
+                idReleasesChannels.push(msgDiscord.channel.id);
+                Promise.all([
+                    db.addReleasesChannel(msgDiscord.channel.id, msgDiscord.guild.id), 
+                    releasesChannels.set(msgDiscord.guild.id, idReleasesChannels)
+                ])
                 msgDiscord.channel.send("Channel added successfully!");
             } else {
-                msgDiscord.channel.send("This channel is already registered!");
+                msgDiscord.channel.send("It's not possible to add more than 10 seperated releases channels!");
             }
         } else {
-            msgDiscord.channel.send("It's not possible to add more than 10 seperated releases channels!");
+            msgDiscord.channel.send("This channel is already registered!");
         }
     } else {
         msgDiscord.reply("You don't have permission to add channels!")
@@ -25,9 +30,16 @@ async function removeChannel(msgDiscord){
     const hasPermission = msgDiscord.member.roles.cache.some(role => role.name === 'New Releases Manager');
 
     if(hasPermission){
-        const channelRemoved = await db.removeReleasesChannel(msgDiscord.channel.id, msgDiscord.guild.id);
-        if(channelRemoved){
-            await db.deleteAllArtistsReleasesChannel(msgDiscord.channel.id);
+        let idReleasesChannels = releasesChannels.get(msgDiscord.guild.id);
+
+        if(idReleasesChannels.includes(msgDiscord.channel.id)){
+            const index = idReleasesChannels.indexOf(msgDiscord.channel.id);
+            idReleasesChannels.splice(index, 1);
+            Promise.all([
+                db.removeReleasesChannel(msgDiscord.channel.id, msgDiscord.guild.id), 
+                db.deleteAllArtistsReleasesChannel(msgDiscord.channel.id),
+                releasesChannels.set(msgDiscord.guild.id, idReleasesChannels)
+            ]);
             msgDiscord.channel.send("Channel deleted successfully!");
         } else {
             msgDiscord.channel.send("This channel is not registered!");
@@ -38,9 +50,13 @@ async function removeChannel(msgDiscord){
 }
 
 async function channelDelete(channel){
-    const channelRemoved = await db.removeReleasesChannel(channel.id);
-    if(channelRemoved){
-        db.deleteAllArtistsReleasesChannel(channel.id);
+    let idReleasesChannels = releasesChannels.get(channel.guild.id);
+
+    if(idReleasesChannels.includes(channel.id)){
+        Promise.all([
+            db.removeReleasesChannel(channel.id),
+            db.deleteAllArtistsReleasesChannel(channel.id)
+        ])
     }
 }
